labs/lab_08_apis/requirements.txt
requests==2.31.0
aiohttp==3.9.1
graphql-core==3.2.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1 